ErrorResultをApiErrorResultに名前変更
