+++ app/src/main/java/com/leoleo/androidgithubsearch/data/paging/GithubRepoPagingSource.kt
@@ -4,7 +4,7 @@ import android.util.Log
 import androidx.paging.PagingSource
 import androidx.paging.PagingState
 import com.leoleo.androidgithubsearch.data.api.GithubService
-import com.leoleo.androidgithubsearch.data.dataOrThrow
+import com.leoleo.androidgithubsearch.data.api.KtorUtil
 import com.leoleo.androidgithubsearch.domain.model.RepositorySummary
 import kotlinx.coroutines.CoroutineDispatcher
 
@@ -30,7 +30,7 @@ class GithubRepoPagingSource(
             val from = pageNumber * size
             val placeholdersEnabled = params.placeholdersEnabled
             val data =
-                dataOrThrow(dispatcher) {
+                KtorUtil.dataOrThrow(dispatcher) {
                     api.searchRepositories(query = query, page = pageNumber)
                 }
             // Since {INIT_PAGE_NO} is the lowest page number, return null to signify no more pages should
