+++ app/src/main/java/com/leoleo/androidgithubsearch/ui/detail/DetailViewModel.kt
@@ -5,7 +5,6 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.leoleo.androidgithubsearch.data.ErrorResult
 import com.leoleo.androidgithubsearch.domain.repository.GithubRepoRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.CoroutineExceptionHandler
